--- app/controllers/users_controller.rb
@@ -1,11 +1,0 @@
-class UsersController < ApplicationController
-  def index
-    binding.pry
-  end
-  def new; end
-  def show; end
-  def destroy; end
-  def update; end
-  def create; end
-  def edit; end
-end
