+++ app/views/tasks/index.html.erb
@@ -15,7 +15,7 @@
         </li>
         <li class="nav-item active">
           <%=link_to "Log out" ,session_path(current_user.id), method: :delete ,class: "nav-link btn btn-outline-primary"%>
-        </li>        
+        </li>
       </ul>
       <%= search_form_for(@q,local: true) do |s|%>
         <%= s.label :name, t('index.index_search')%>
@@ -37,26 +37,24 @@
       <th colspan = 4 scope = "col"><%= t('index.status')%></th>
     </tr>
     <%@tasks.each do |task|%>
-      <% if task.user_id == current_user.id%>
-        <tr>
-          <td><p class = index_user_name><%= task.user.name %>&nbsp;</p></td>
-          <td><p class = index_name><%= task.name%>&nbsp;</p></td>
-          <td><p class = index_detail><%= task.detail%>&nbsp;</p></td>
-          <td><p class = index_deadline><%= task.deadline%>&nbsp;</p></td>
-          <td><p class = index_priority>
-            <% if task.priority %>
-              <% if task.priority.split.last == "高" %>
-                <p class = task_priority_color><%= task.priority.split.last%></p>
-              <% else %>
-                <%= task.priority.split.last %>
-              <% end %>
-            <% end %>  
-          <td><p class = index_status><%= task.status%>&nbsp;</p></td>
-          <td class="btn-link"><p class = index_details_button><%= link_to t('index.details'),task_path(task.id)%>&nbsp;</p></td>
-          <td class="btn-link"><%= link_to t('index.edit'),edit_task_path(task.id)%>&nbsp;</td>
-          <td class="btn-link"><%= link_to t('index.destroy'),task_path(task.id),method: :delete,data:{confirm: "削除してもよろしいですか"}%>&nbsp;</td>
-        </tr>
-      <% end %>        
+      <tr>
+        <td><p class = index_user_name><%= task.user.name %>&nbsp;</p></td>
+        <td><p class = index_name><%= task.name%>&nbsp;</p></td>
+        <td><p class = index_detail><%= task.detail%>&nbsp;</p></td>
+        <td><p class = index_deadline><%= task.deadline%>&nbsp;</p></td>
+        <td><p class = index_priority>
+          <% if task.priority %>
+            <% if task.priority.split.last == "高" %>
+              <p class = task_priority_color><%= task.priority.split.last%></p>
+            <% else %>
+              <%= task.priority.split.last %>
+            <% end %>
+          <% end %>  
+        <td><p class = index_status><%= task.status%>&nbsp;</p></td>
+        <td class="btn-link"><p class = index_details_button><%= link_to t('index.details'),task_path(task.id)%>&nbsp;</p></td>
+        <td class="btn-link"><%= link_to t('index.edit'),edit_task_path(task.id)%>&nbsp;</td>
+        <td class="btn-link"><%= link_to t('index.destroy'),task_path(task.id),method: :delete,data:{confirm: "削除してもよろしいですか"}%>&nbsp;</td>
+      </tr>
     <% end %>
   </table>
   <%=link_to t('index.create'),new_task_path%>
